requirements.txt
absl-py==2.2.2
amqp==5.3.1
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
babel==2.17.0
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
debugpy==1.8.12
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
django-rest-knox==5.0.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.1.0
flatbuffers==25.2.10
fqdn==1.5.1
gast==0.6.0
git-filter-repo==2.47.0
google-pasta==0.2.0
gunicorn==20.1.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
ipykernel==6.29.5
ipython==8.32.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.9.2
kombu==5.4.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
ml_dtypes==0.5.1
msgpack==1.1.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy>=1.21,<2.2
opt_einsum==3.4.0
optree==0.15.0
overrides==7.7.0
onedrivedownloader==1.0.0  
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1

PyYAML==6.0.2
pyzmq==26.2.1
redis==5.1.1
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.21.0
Send2Trash==1.8.3
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.2
stack-data==0.6.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-cpu==2.19.0 
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
terminado==0.18.1
tinycss2==1.4.0
tomli==2.1.0
tornado==6.4.2
traitlets==5.14.3
Twisted==24.10.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.27.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
zope.interface==7.1.1
